refactor(botonic-react): grop all components used only in webchat-input-panel inside his folder
