--- packages/botonic-react/src/webchat/textarea/styles.ts
@@ -1,7 +1,0 @@
-import styled from 'styled-components'
-
-export const TextAreaContainer = styled.div`
-  display: flex;
-  flex: 1 1 auto;
-  align-items: center;
-`
