+++ packages/botonic-react/src/webchat/webchat-input-panel/emoji-picker.jsx
@@ -1,13 +1,13 @@
 import Picker from 'emoji-picker-react'
 import React, { useContext } from 'react'
-import styled from 'styled-components'
 
 import LogoEmoji from '../../assets/emojiButton.svg'
 import { ROLES, WEBCHAT } from '../../constants'
 import { WebchatContext } from '../../contexts'
+import { Icon } from '../components/common'
 import { ConditionalAnimation } from '../components/conditional-animation'
 import { useComponentVisible } from '../hooks'
-import { Icon } from './common'
+import { OpenedEmojiPickerContainer } from './styles'
 
 export const EmojiPicker = ({ enableEmojiPicker, onClick }) => {
   const { getThemeProperty } = useContext(WebchatContext)
@@ -50,20 +50,12 @@ export const EmojiPicker = ({ enableEmojiPicker, onClick }) => {
   )
 }
 
-const Container = styled.div`
-  display: flex;
-  justify-content: flex-end;
-  position: absolute;
-  right: 3px;
-  top: -324px;
-`
-
 export const OpenedEmojiPicker = props => {
   const { ref, isComponentVisible } = useComponentVisible(true, props.onClick)
   return (
     <div ref={ref}>
       {isComponentVisible && (
-        <Container role={ROLES.EMOJI_PICKER}>
+        <OpenedEmojiPickerContainer role={ROLES.EMOJI_PICKER}>
           <Picker
             width='100%'
             height='19rem'
@@ -71,7 +63,7 @@ export const OpenedEmojiPicker = props => {
             onEmojiClick={props.onEmojiClick}
             disableAutoFocus={true}
           />
-        </Container>
+        </OpenedEmojiPickerContainer>
       )}
     </div>
   )
