+++ packages/botonic-react/src/webchat/webchat-input-panel/persistent-menu.jsx
@@ -1,0 +1,33 @@
+import React, { useContext } from 'react'
+
+import LogoMenu from '../../assets/menuButton.svg'
+import { ROLES, WEBCHAT } from '../../constants'
+import { WebchatContext } from '../../contexts'
+import { Icon } from '../components/common'
+import { ConditionalAnimation } from '../components/conditional-animation'
+
+export const PersistentMenu = ({ onClick, persistentMenu }) => {
+  const { getThemeProperty } = useContext(WebchatContext)
+
+  const persistentMenuOptions = getThemeProperty(
+    WEBCHAT.CUSTOM_PROPERTIES.persistentMenu,
+    persistentMenu
+  )
+
+  const CustomMenuButton = getThemeProperty(
+    WEBCHAT.CUSTOM_PROPERTIES.customMenuButton,
+    undefined
+  )
+
+  return (
+    <>
+      {persistentMenuOptions ? (
+        <ConditionalAnimation>
+          <div role={ROLES.PERSISTENT_MENU_ICON} onClick={onClick}>
+            {CustomMenuButton ? <CustomMenuButton /> : <Icon src={LogoMenu} />}
+          </div>
+        </ConditionalAnimation>
+      ) : null}
+    </>
+  )
+}
