+++ packages/botonic-react/src/webchat/webchat-input-panel/send-button.jsx
@@ -3,8 +3,8 @@ import React, { useContext } from 'react'
 import SendButtonIcon from '../../assets/send-button.svg'
 import { ROLES, WEBCHAT } from '../../constants'
 import { WebchatContext } from '../../contexts'
+import { Icon } from '../components/common'
 import { ConditionalAnimation } from '../components/conditional-animation'
-import { Icon } from './common'
 
 export const SendButton = ({ onClick }) => {
   const { getThemeProperty } = useContext(WebchatContext)
